Fix incorrect "deque" term use internally
Referring to the combined front and back marker tracking stacks as a
"deque" internally is incorrect and misleading, as the two stacks
operate independently of each other for the most part (aside from
sharing memory). They should be referred to as separate stacks to avoid
any confusion when looking through the code.