Dereference data in `Allocation` debug formatting
The `fmt::Debug` implementation for `Allocation` included the allocation
data in its output, but this was changed to only output its raw pointer
value when the `'t` lifetime parameter was removed. The data itself is
more useful in this context, and the raw pointer is never null, so the
original functionality has been restored.