Replace `AsMutSlice` with `IntoSliceAllocation`
`IntoSliceAllocation` more explicitly marks types that can be converted
into slice allocations both in its name as well as by moving
implementations of the trait to just the `Allocator` type instead of
general scalar, array, or slice types. It also adds functionality for
explicitly coercing an allocation into a slice allocation. It is also
marked as `unsafe` to better signify that it is not intended to be
implemented freely by other miscellaneous types.