Use consistent slice types with "extend" methods
This modifies the `Marker::extend_clone()` and `Marker::extend_copy()`
(and the related `MarkerFront` and `MarkerBack` methods) to take the
same types of slice parameters as the corresponding slice allocation
methods (`allocate_slice_copy()` and `allocate_slice_clone()`). This not
only improves interface consistency, but also allows for DSTs that wrap
slice types such as `str` to be used with the extend/append/prepend
methods. The extend method implementations now also make use of the
corresponding slice allocation methods to reduce redundant code.