Generalize slice inputs for `Marker` methods
This adds the `AsSliceLike` trait for returning `SliceLike` references
from potentially non-`SliceLike` types. The `Marker` methods for
allocating slices, extending allocations, and concatenating slices that
clone or copy from their inputs have been modified to make use of this
trait to allow for a wider range of inputs, such as scalars and arrays.