Allow flexible slice input in all `Marker` methods
This reworks all `Marker` methods that take slice input to allow for
more flexible input types by way of several new traits. `SliceSource`
replaces `AsSliceLike` and `OwnedSlice` for providing slice data from
potentially non-slice types, with the `SliceMoveSource` subtrait
providing functionality for taking ownership of the contents of a slice
(safer than the way in which the `OwnedSlice` trait was used).
`SliceSourceCollection` provides an abstraction layer for collections
(namely tuples and arrays) of `SliceSource` objects, with the
`SliceMoveSourceCollection` subtrait allowing for taking ownership of
the contents of each slice provided by the collection.