Remove automatic coercion of scalars into slices
Automatic coercion of scalars into slices when used as arguments for
various `Marker` functions such as `Marker::allocate_slice()` often
results in ambiguous type inference when passing arrays or slice
references, as each of those could be interpreted as either a slice of
their contents or a single-element slice containing either the array or
reference itself, requiring explicit type annotations by the user. Since
the use of scalars is more of an edge case, it is arguably better to
require the user to explicitly convert such types into arrays (by
enclosing with square brackets) or slices (using something akin to
`std::slice::from_raw_parts()` or the `ref_slice` crate), allowing the
compiler to automatically coerce arrays and slice references into slices
of their contents. This also helps us push to reducing compile times and
metadata overhead for this crate (additional optimizations are planned
that would require something like this to be done first).