Add `Array` trait, simplify trait implementations
The `Array` trait is used to allow generic implementations of other
traits for static arrays instead of the existing method of implementing
the traits for each specific array type, helping reduce compile times
and metadata bloat.

While this works out nicely for implementing traits for the static array
types themselves, attempting to implement traits for array references
and boxed arrays results in compiler errors due to the possibility of
downstream crates implementing `Array` for references or boxed instances
of types they may define. Unfortunately, we cannot prevent downstream
crates from doing so with current versions of Rust, so we will no longer
implement such traits for array references and boxed arrays. This
shouldn't present too much of an issue, as such types can easily be
converted into slice references or boxed slices, respectively.

Note that `IntoSliceAllocation` is still implemented explicitly for each
array type at this time, as creating a blanket implementation based on
the `Array` trait conflicts with the implementation for scalar type
allocations.