Implement `Tracking` for aligned `Buffer`s only
Implementing `Tracking` for all `Buffer` implementations forced us to
use unaligned reads and writes of tracking values in order to avoid
possible issues with architectures that don't allow direct unaligned
access. This change restricts the default `Tracking` implementation to
`Buffer` types for which proper alignment can be guaranteed (by way of
the new `SizeAlignedByteData` and `SizeAlignedBuffer` traits) and
replaces the unaligned access calls with direct memory reads and writes.