Declare `into_mut_slice_like_ptr()` as unsafe
Implementations of `IntoMutSliceLikePtr::into_mut_slice_like_ptr()` may
require reading the data pointed to by the given pointer, so they should
be marked as `unsafe` to indicate this (this was previously noted, but
could not be addressed in the 1.x releases without potentially breaking
compatibility). This change also allows us to remove attributes for
ignoring `clippy` lint errors that were generated with the existing
trait method implementations.